models/where_predictor.py
# models/where_predictor.py

import torch
import torch.nn as nn
import torch.nn.functional as F

class WherePredictor(nn.Module):
    def __init__(self, hidden_dim, glove_embeddings):
        super(WherePredictor, self).__init__()
        self.glove_embeddings = glove_embeddings
        self.hidden_dim = hidden_dim
        self.pointer_network = nn.LSTM(hidden_dim, hidden_dim, batch_first=True)  # Example implementation
        self.W_ptr = nn.Linear(hidden_dim, hidden_dim)
        self.U_ptr = nn.Linear(hidden_dim, hidden_dim)
        self.V_ptr = nn.Linear(hidden_dim, hidden_dim)

    def scalar_attention(self, scores, hidden_states):
        scores = scores.permute(0, 2, 1)  # [batch_size, hidden_dim, seq_len]
        attention_weights = torch.bmm(scores, hidden_states.unsqueeze(2)).squeeze(2)  # [batch_size, seq_len]
        return F.softmax(attention_weights, dim=-1)

    def forward(self, input_encodings):
        output, (hn, cn) = self.pointer_network(input_encodings)
        hn_last = hn[-1]
        scores = self.scalar_attention(output, hn_last)
        return scores

    def compute_reward(self, generated_query, ground_truth):
        if not self.is_valid_sql(generated_query):
            return -2
        elif generated_query != ground_truth:
            return -1
        else:
            return 1

    def is_valid_sql(self, query):
        # Implement a proper SQL validation here
        return True

    def train_step(self, input_encodings, generated_query, ground_truth):
        scores = self.forward(input_encodings)
        reward = self.compute_reward(generated_query, ground_truth)
        loss = -reward * torch.sum(scores)
        return loss
